Split ListService
